pepega1.0

+++ restapi/application_views.py
@@ -2,7 +2,7 @@ from rest_framework.response import Response
 from rest_framework.decorators import api_view
 from .serializers import JobSerializer,application_Serializer,UserLoginSerializer
 from .models import job,User,application,spez
-from .views import register
+from .views import register,authAdmin,authuser
 
 @api_view(['POST'])
 def register_Application(request):
@@ -39,17 +39,21 @@ def hireability_score(request):
 
 @api_view(['POST'])
 def update_Application(request):
-    ass = application_Serializer(data=request.data)
+    user = authAdmin(request)
+    if user:
+        ass = application_Serializer(data=request.data)
     # obj = application.objects.filter(id=ass.id)
     # obj.postal=ass.data.get('postal')
     # obj.city=ass.data.get('city')
     # obj.pincode=ass.data.get('pincode')
     # obj.mob_num=ass.data.get('mob_num')
     # obj.state=ass.data.get('state')
-    if ass.is_valid():
-        obj = ass.save()
+        if ass.is_valid():
+            obj = ass.save()
+        else:
+            return Response({"bad":"response"})
     else:
-        return Response({"bad":"response"})
+        return Response({"bad":"user"})
     # user.application = obj
     # obj.spez_Req=spez
     # jb = job.objects.filter(id=request.data.get("job_id"))
@@ -65,19 +69,41 @@ def update_Application(request):
 
 @api_view(['POST'])
 def delete_app(request):
-    job.objects.filter(id=request.data.get("id")).delete()
+    user = authAdmin(request)
+    if user:
+        job.objects.filter(id=request.data.get("id")).delete()
+
+    else:
+        return Response({"bad":"user"})
 
 @api_view(['POST'])
 def nextRnd(request):
-    jb=job.objects.filter(id=request.data.get("id"))
-    jb.round+=1
-    jb.save()
+    user = authAdmin(request)
+    if user:
+        jb=job.objects.filter(id=request.data.get("id"))
+        jb.round+=1
+        jb.save()
+    else:
+        return Response({"bad":"user"})
+    
 
 @api_view(['POST'])
 def schedule(request):
-    jb=job.objects.filter(id=request.data.get("id"))
-    jb.schedule = request.data.get("datetime")
-    jb.save()
+    user = authAdmin(request)
+    if user:
+        jb=job.objects.filter(id=request.data.get("id"))
+        jb.schedule = request.data.get("datetime")
+        jb.save()
+    else:
+        return Response({"bad":"user"})
+    
 
 @api_view(['GET'])
 def get_details(request):
+    user = authuser(request)
+    if user:
+        ap = application.objects.filter(user=user)
+        return Response(application_Serializer(ap).data)
+    else:
+        Response({"bad":"auth"})
+
